Moved hash data to zip module
