+++ modules/i3s/src/lib/parsers/parse-slpk/parse-slpk.ts
@@ -1,14 +1,12 @@
-import md5 from 'md5';
 import {
   parseZipCDFileHeader,
   cdSignature as cdHeaderSignature,
-  parseEoCDRecord,
   FileProvider,
   parseZipLocalFileHeader,
   searchFromTheEnd,
   HashElement,
-  compareHashes,
-  parseHashFile
+  parseHashFile,
+  generateHashInfo
 } from '@loaders.gl/zip';
 import {SLPKArchive} from './slpk-archieve';
 
@@ -51,27 +49,4 @@ export const parseSLPK = async (
   }
 
   return new SLPKArchive(fileProvider, hashData);
-};
-
-/**
- * generates hash info from central directory
- * @param fileProvider - provider of the archive
- * @returns ready to use hash info
- */
-const generateHashInfo = async (fileProvider: FileProvider): Promise<HashElement[]> => {
-  const {cdStartOffset} = await parseEoCDRecord(fileProvider);
-  let cdHeader = await parseZipCDFileHeader(cdStartOffset, fileProvider);
-  const hashInfo: HashElement[] = [];
-  while (cdHeader) {
-    hashInfo.push({
-      hash: Buffer.from(md5(cdHeader.fileName.split('\\').join('/').toLocaleLowerCase()), 'hex'),
-      offset: cdHeader.localHeaderOffset
-    });
-    cdHeader = await parseZipCDFileHeader(
-      cdHeader.extraOffset + BigInt(cdHeader.extraFieldLength),
-      fileProvider
-    );
-  }
-  hashInfo.sort((a, b) => compareHashes(a.hash, b.hash));
-  return hashInfo;
 };
